+++ models/portfolio.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Portfolio = sequelize.define('Portfolio', {
     tenant: DataTypes.STRING,
     buildingSize: DataTypes.INTEGER,
-    lotSize: DataTypes.INTEGER,
+    lotSize: DataTypes.DECIMAL(10,2),
     type: { 
         type: DataTypes.ENUM,
         values: ['Commercial', 'Vacant', 'Mixed Use']
